src/python/requirements.txt
# Python dependencies for WhisperSermons sermon processing
# These are installed on first run via the setup wizard

# Core transcription - MLX-accelerated Whisper for Apple Silicon
mlx-whisper>=0.4.0

# Unified embedding model for paragraph segmentation and tag extraction
# Replaces torch + sentence-transformers + keybert (~2GB dependency reduction)
# Uses EmbeddingGemma-300m-4bit via native MLX GPU acceleration
mlx-embeddings>=0.0.5

# Audio metadata extraction
mutagen>=1.47.0

# HTTP requests (for optional online Bible API fallback)
requests>=2.31.0

# Numerical computing
numpy>=1.24.0
